Mirror of EA WebKit 16.4.2.0.0 from https://gpl.ea.com/eawebkit.html

+++ WebCore/platform/network/ea/CookieManagerEA.cpp
@@ -40,14 +40,14 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "DateEA.h"
 #include "URL.h"
 
-#include <EAWebkit/EAWebKitTransport.h>
+#include <EAWebKit/EAWebKitTransport.h>
 #include <EAWebKit/EAWebKitFileSystem.h>
 #include <internal/include/EAWebKit_p.h>
 #include <internal/include/EAWebKitAssert.h>
 #include <internal/include/EAWebKitString.h>
 #include <internal/include/EAWebKitNewDelete.h>
 #include <EASTL/sort.h>
-#include <eastl/fixed_vector.h>
+#include <EASTL/fixed_vector.h>
 
 #include <EAIO/PathString.h>
 
@@ -1131,7 +1131,7 @@ namespace EA
                             }
                             else
                             {
-                                EAW_ASSERT_FORMATTED(false, "CookieManager: Not enough buffer for ports. Required:%u, Available:%d\n", portsStr.length(), portsBufferLength);
+                                EAW_ASSERT_FORMATTED(false, "CookieManager: Not enough buffer for ports. Required:%u, Available:%d\n", portsStr.length(), (int)portsBufferLength);
                             }
 
                         }
